refactor(core)!: one score scale for a search candidate and an identity
SearchCandidate.score was MusicBrainz's own 0-100 relevance leaking onto the
published surface, which every other score on it gives as a Float on 0.0-1.0.
It becomes matchScore: Float, normalised and clamped at the mapper, and the
constructor is reordered required-first. IdentityResolution.matchScore and the
three profiles' identityMatchScore follow it onto the same scale.

+++ demo-web/src/main/kotlin/com/landofoz/musicmeta/demoweb/ApiModels.kt
@@ -280,7 +280,8 @@ data class SearchResponse(val candidates: List<CandidateHit>)
 
 /**
  * One `SearchCandidate` on the wire. [mbid] is the candidate's `identifiers.musicBrainzId`: when
- * it is present the page enriches by identifier instead of by name.
+ * it is present the page enriches by identifier instead of by name. [score] carries the
+ * candidate's `matchScore`, 0.0-1.0.
  */
 @Serializable
 data class CandidateHit(
@@ -288,7 +289,7 @@ data class CandidateHit(
     val artist: String? = null,
     val year: String? = null,
     val releaseType: String? = null,
-    val score: Int,
+    val score: Float,
     val thumbnailUrl: String? = null,
     val disambiguation: String? = null,
     val mbid: String? = null,
